+++ test/index.test.js
@@ -32,3 +32,4 @@ const isFalsity = require("./cases/isFalsity.test.js");
 const isBigInt = require("./cases/isBigInt.test.js");
 const isPrimitive = require("./cases/isPrimitive.test.js");
 const isWeakMap = require("./cases/isWeakMap.test.js");
+const isEmptyArray = require("./cases/isEmptyArray.test.js");